requirements.txt
django
djangorestframework
pyJWT
djangorestframework-simplejwt
Pillow